feat + fix: Add all routes controllers to routes/index.ts

+++ src/controllers/funcionario/DeleteById.ts
@@ -5,7 +5,7 @@ import { FuncionarioProvider } from "../../server/database/providers/funcionario
 import { StatusCodes } from "http-status-codes";
 
 interface IParamProps {
-  id_funcionario: number;
+  id_funcionario?: number;
 }
 
 export const deleteByIdValidation = validation((getSchema) => ({
@@ -17,6 +17,12 @@ export const deleteByIdValidation = validation((getSchema) => ({
 }));
 
 export const deleteById = async (req: Request<IParamProps>, res: Response) => {
+  if (!req.params.id_funcionario) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      errors: { default: "O parâmetro 'id_cargo' é obrigatório" },
+    });
+  }
+
   const result = await FuncionarioProvider.deleteById(
     req.params.id_funcionario
   );
